+++ packages/message/tests/test-Queue.js
@@ -3,31 +3,12 @@
 const test = require('ava');
 const cryptoRandomString = require('crypto-random-string');
 const {
-  getQueueName,
   getQueueUrl,
-  getQueueNameByUrl,
   getMaximumExecutions,
   hasQueueAndExecutionLimit,
 } = require('../Queue');
 
 const randomId = (prefix) => `${prefix}${cryptoRandomString({ length: 10 })}`;
-
-test('getQueueName returns correct queue name', (t) => {
-  const queueName = randomId('queue');
-  const message = {
-    cumulus_meta: {
-      queueName,
-    },
-  };
-  t.is(getQueueName(message), queueName);
-});
-
-test('getQueueName throws an error if cumulus_meta.queueName is not set in message', (t) => {
-  const message = {
-    cumulus_meta: {},
-  };
-  t.throws(() => getQueueName(message));
-});
 
 test('getQueueUrl returns correct queue URL', (t) => {
   const queueUrl = randomId('queue');
@@ -44,27 +25,6 @@ test('getQueueUrl throws error if queue URL cannot be found', (t) => {
     cumulus_meta: {},
   };
   t.throws(() => getQueueUrl(message));
-});
-
-test('getQueueNameByUrl returns correct value', (t) => {
-  const queueName = randomId('queueName');
-  const queueUrl = randomId('queueUrl');
-  const testMessage = {
-    meta: {
-      queues: {
-        [queueName]: queueUrl,
-      },
-    },
-  };
-
-  let queueNameResult = getQueueNameByUrl(testMessage, queueUrl);
-  t.is(queueNameResult, queueName);
-
-  queueNameResult = getQueueNameByUrl(testMessage, 'fake-value');
-  t.is(queueNameResult, undefined);
-
-  queueNameResult = getQueueNameByUrl({}, 'queueUrl');
-  t.is(queueNameResult, undefined);
 });
 
 test('getMaximumExecutions returns correct value', (t) => {